Feature - Discount Integration (#93)
* :construction: wip make payment

* :sparkles: display order summary

* :sparkles: integrate payment history

* :lock: guard pay order with no order items

* :sparkles: show notification on pay order

* :bug: fix duplicate notifications

* :wrench: auth service config

* add discounts to gateway

* wip displaying of discounts

* update model and display discounts

* :hammer: handle closing of payment dialog

* :safety_vest: add payment amount validation

* :recycle: refactor

* :lipstick: add transitions for splitview pane

* :recycle: create separate style file for side nav bar

* :adhesive_bandage: fix namespace import

* :lipstick: remove unnecessary border of side nav bar

* :lipstick: change width ratio of order cart panel and wait staff page

* :lipstick: create file for tab strip item stylings

* add apply discount and wip in remove discount

* :art: create info dialog component

* :lipstick: use infodialog instead of built in

* :lipstick: add title to info dialogs

* display discount on cart panel and order summary

* :art: make all comparisons to string

* :lipstick: show remove when discount is applied

* :art: create global variable for viewmodel

* wip fetch discount order

* handle DiscountOrder on LoadOrder

* :construction: add body to delete request

* update DeleteAsync method

* update discount service config

* :art: show applied discount upon opening discount dialog

* :fire: remove unused variable

* wip display discounts on OrderCartPanel

* :bug: fix null viewModel

* :art: disable order discount button when line item discount is applied

* :recycle: use message instead of variable binding when disabling order discount button

* :recycle: :art: prevent order and line item discount from being applied simultaneously

* :bug: fix individual line item discount not displaying which is applied

* :bug: fix apply order discount button still disabled when order items is cleared

* wip disable buttons on discount apply

* wip apply discount buttons

* :sparkles: show discount amount in line item

* :lipstick: minor ui adjustments to info dialog

* :sparkles: add close button to discount dialog

* :lipstick: use same close button from info dialog

* :art: make close dialog code single line

* :art: :lipstick: create custom component for two fold dialog and modify ui

* :lipstick: minor ui adjustments in info dialog

* :art: use confirmation dialog instead

* :lipstick: minor ui adjustments to dialogs

* :lipstick: allow text wrapping of dialog title

* handle discount ui validations

* :lipstick: adjust category dialog ui

* :lipstick: show status when line item discount is applied

* :lipstick: fix spacing larger on left side of waitstaff page

* add validation for applying multiple discount

* wip refactor OrderTransaction model

* :lipstick: make discounted discount text dimmer

* :children_crossing: display discount name in line item

* add DiscountOrderLineitemDiscount model

* update AppliedDiscountId

* fix displaying of discounted amount in line item

* :children_crossing: show applied order discount status

* prevent applying discount when payment is on going

* :children_crossing: show discount name in line item discount button

* :hammer: allow zero amount on payment, clear discounts on new order

* :speech_balloon: update negative amount validation message

* :lipstick: fix large spacing in products page

* :lipstick: use info dialog instead of single action dialog of avaloniadialogs

* :hammer: clear discounts properly

* :adhesive_bandage: fix ui bug displaying two discounts being applied

* :lipstick: change discount error dialog title and message

* ♻️

* ♻️ refactor update discount

* ♻️ update remove line item for order level

* remove text

* :bug: Bug | Save Customer (#92)

* :construction: wip update customer name validation

* :hammer: add name validation on product selection, remove creating order on customer field

* :hammer: move validation on create order only

* :goal_net: handle command errors

* :safety_vest: add empty customer name validation

* update discount dialog title

* :lock: change admin password

* :lock: change staff names

---------

+++ UI/Handlers/HttpHandler.cs
@@ -7,6 +7,7 @@ using pos.Handlers.Interfaces;
 using pos.Api;
 using Microsoft.Extensions.DependencyInjection;
 using Pos;
+using System.Text;
 
 namespace pos.Handlers
 {
@@ -68,10 +69,19 @@ namespace pos.Handlers
             return await _client.PutAsJsonAsync(endpoint, putData);
         }
 
-        public async Task<HttpResponseMessage> DeleteAsync(string endpoint, object? caller = null)
+        public async Task<HttpResponseMessage> DeleteAsync(string endpoint, object? deleteData = null, object? caller = null)
         {
             await AddAuthHeaderAsync(caller);
-            return await _client.DeleteAsync(endpoint);
+            if (deleteData != null)
+            {
+                using var request = new HttpRequestMessage(HttpMethod.Delete, endpoint)
+                {
+                    Content = new StringContent(JsonSerializer.Serialize(deleteData), Encoding.UTF8, "application/json")
+                };
+                return await _client.SendAsync(request);
+            }
+            else
+                return await _client.DeleteAsync(endpoint);
         }
 
         public async Task<T> ReadJsonResponseAsync<T>(HttpResponseMessage response)
